requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
decouple==0.0.7
dj-database-url==3.0.0
Django==5.2.3
easydict==1.13
et_xmlfile==2.0.0
flatbuffers==25.2.10
fonttools==4.58.3
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
gspread-formatting==1.2.1
gunicorn==23.0.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
oauth2client==4.1.3
oauthlib==3.2.2
onnx==1.18.0
onnxruntime==1.22.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prettytable==3.16.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.6
pydantic_core==2.33.2
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
simsimd==6.4.9
six==1.17.0
sqlparse==0.5.3
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.11
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
whitenoise==6.9.0
